Refactor Qwen3 14B decode into fused inline path
Split the Qwen3 14B decode implementation into shared config, a single-layer inline decode_layer kernel, and a full-forward decode_fwd wrapper.

Move final RMSNorm and LM head into decode_layer so the fused layer kernel owns logits production, and have decode_fwd call decode_layer directly across num_layers_actual.

Remove the old qwen3_14b_decode, qwen3_14b_decode_full, final_rms, and lm_head files, and update the golden/run_jit wiring for the flattened harness API.
